move sensitive_string
